requirements.txt
Django==1.4
django-piston==0.2.2
gunicorn==0.14.2
hiredis==0.1.1
mongoengine==0.6.6
pymongo==2.1.1
redis==2.4.11
wsgiref==0.1.2
